requirements.txt
# Core Dependencies
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
Pillow>=10.0.0
opencv-python>=4.8.0

# Machine Learning
tensorflow>=2.13.0
scikit-learn>=1.3.0

# For Face Detection
mediapipe>=0.10.0

# Data Handling
kaggle>=1.5.16

# TFLite and Edge TPU (for deployment phase)
# Note: These will be installed on Raspberry Pi later
# tflite-runtime>=2.13.0
# pycoral

# Utilities
tqdm>=4.65.0
jupyter>=1.0.0
ipykernel>=6.25.0

# Testing
pytest>=7.4.0

